database connection and plot
the database connection is made through the global.r script and the option of the request (WHERE) are made on the server.r function wich also draw the plot regarding the data on the database.

Theses plots are active whenever a button is pressed and depend of the Email selected in the dropdown.