server.R: Change Default E-mail based on URL Param
Ideally we want to segway our students straight to
their data, rather than describing the path they
should take in instructions. This requires that
we can filter the data based on a predefined URL
parameter. The parameter in question is email="".
If no parameter is specified, we default to the
first e-mail available in the list.