server.R: Fix PIDChooser bugs
PIDs are stored in a named vector with a PID_name
and an underlying index. This is required when different
e-mail addresses use the same participant numbers.
However, when querying as URL parameter from Unity,
we are only aware of the PID_name and not the underlying
index, which is actually what is used to refer to the PID.

This commit fixes this and removes bugs related to filtering
the data. It also makes sure that only the PIDs related to
the given tabs are shown each at a time.